pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "commitment-os"
version = "0.1.0"
description = "CommitmentOS: the first RL environment that trains temporal commitment coherence in LLMs"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Jayant Aggarwal"},
]

dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
inference = [
    "openai>=1.0.0",
    "requests>=2.31.0",
]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
]
training = [
    "trl>=0.14.0",
    "transformers>=4.45.0",
    "torch>=2.0.0",
    "peft>=0.14.0",
    "datasets>=3.0.0",
]
